renaming updateDetailedItem to reloadDetailedItem;
