+++ ui/app/src/state/epics/content.ts
@@ -26,15 +26,18 @@ import {
   fetchUserPermissions,
   fetchUserPermissionsComplete,
   fetchUserPermissionsFailed,
-  updateDetailedItem
+  reloadDetailedItem
 } from '../actions/content';
 import { catchAjaxError } from '../../utils/ajax';
-import { fetchQuickCreateList, getDetailedItem } from '../../services/content';
+import { duplicate, fetchQuickCreateList, getDetailedItem } from '../../services/content';
 import StandardAction from '../../models/StandardAction';
 import GlobalState from '../../models/GlobalState';
 import { GUEST_CHECK_IN } from '../actions/preview';
 import { getUserPermissions } from '../../services/security';
 import { NEVER } from 'rxjs';
+import { assetDuplicate, itemDuplicate } from '../actions/misc';
+import { showCodeEditorDialog, showEditDialog } from '../actions/dialogs';
+import { isEditableAsset } from '../../utils/content';
 
 export default [
   // region Quick Create
@@ -74,10 +77,10 @@ export default [
   // region Items fetchDetailedItem
   (action$: ActionsObservable<StandardAction>, state$: StateObservable<GlobalState>) =>
     action$.pipe(
-      ofType(fetchDetailedItem.type, updateDetailedItem.type),
+      ofType(fetchDetailedItem.type, reloadDetailedItem.type),
       withLatestFrom(state$),
       switchMap(([{ payload, type }, state]) => {
-          if (type !== updateDetailedItem.type && state.content.items.byId?.[payload.path]) {
+          if (type !== reloadDetailedItem.type && state.content.items.byId?.[payload.path]) {
             return NEVER;
           } else {
             return getDetailedItem(state.sites.active, payload.path).pipe(
@@ -87,6 +90,42 @@ export default [
           }
         }
       )
+    ),
+  // endregion
+  // region itemDuplicate
+  (action$, state$: StateObservable<GlobalState>) =>
+    action$.pipe(
+      ofType(itemDuplicate.type),
+      withLatestFrom(state$),
+      switchMap(([{ payload }, state]) => {
+        return duplicate(state.sites.active, payload.path).pipe(
+          map((path) => {
+            const src = `${state.env.authoringBase}/legacy/form?site=${state.sites.active}&path=${path}&type=form`;
+            debugger;
+            return showEditDialog({ src, onSaveSuccess: payload.onSuccess });
+          })
+        );
+      })
+    ),
+  // endregion
+  // region assetDuplicate
+  (action$, state$: StateObservable<GlobalState>) =>
+    action$.pipe(
+      ofType(assetDuplicate.type),
+      withLatestFrom(state$),
+      switchMap(([{ payload }, state]) => {
+        return duplicate(state.sites.active, payload.path).pipe(
+          map((path) => {
+            const editableAsset = isEditableAsset(payload.path);
+            if (editableAsset) {
+              const src = `${state.env.authoringBase}/legacy/form?site=${state.sites.active}&path=${path}&type=asset`;
+              return showCodeEditorDialog({ src, onSuccess: payload.onSuccess });
+            } else {
+              return payload.onSuccess;
+            }
+          })
+        );
+      })
     )
   // endregion
 ];
