+++ ui/app/src/state/epics/misc.ts
@@ -14,25 +14,22 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { Epic, ofType, StateObservable } from 'redux-observable';
+import { Epic, ofType } from 'redux-observable';
 import { map, switchMap, withLatestFrom } from 'rxjs/operators';
 import { NEVER, Observable } from 'rxjs';
 import GlobalState from '../../models/GlobalState';
 import {
-  assetDuplicate,
   changeContentType as changeContentTypeAction,
-  editTemplate as editTemplateAction,
-  itemDuplicate
+  editTemplate as editTemplateAction
 } from '../actions/misc';
 import queryString from 'query-string';
-import { changeContentType, duplicate, fetchWorkflowAffectedItems } from '../../services/content';
+import { changeContentType, fetchWorkflowAffectedItems } from '../../services/content';
 import {
   showCodeEditorDialog,
   showEditDialog,
   showWorkflowCancellationDialog
 } from '../actions/dialogs';
 import { pathNavigatorItemActionSuccess } from '../actions/pathNavigator';
-import { isEditableAsset } from '../../utils/content';
 
 const changeTemplate: Epic = (action$, state$: Observable<GlobalState>) => action$.pipe(
   ofType(changeContentTypeAction.type),
@@ -71,43 +68,7 @@ const editTemplate: Epic = (action$, state$: Observable<GlobalState>) => action$
   })
 );
 
-const duplicateItem: Epic = (action$, state$: StateObservable<GlobalState>) =>
-  action$.pipe(
-    ofType(itemDuplicate.type),
-    withLatestFrom(state$),
-    switchMap(([{ payload }, state]) => {
-      return duplicate(state.sites.active, payload.path).pipe(
-        map((path) => {
-          const src = `${state.env.authoringBase}/legacy/form?site=${state.sites.active}&path=${path}&type=form`;
-          debugger;
-          return showEditDialog({ src, onSaveSuccess: payload.onSuccess });
-        })
-      );
-    })
-  );
-
-const duplicateAsset: Epic = (action$, state$: StateObservable<GlobalState>) =>
-  action$.pipe(
-    ofType(assetDuplicate.type),
-    withLatestFrom(state$),
-    switchMap(([{ payload }, state]) => {
-      return duplicate(state.sites.active, payload.path).pipe(
-        map((path) => {
-          const editableAsset = isEditableAsset(payload.path);
-          if (editableAsset) {
-            const src = `${state.env.authoringBase}/legacy/form?site=${state.sites.active}&path=${path}&type=asset`;
-            return showCodeEditorDialog({ src, onSuccess: payload.onSuccess });
-          } else {
-            return payload.onSuccess;
-          }
-        })
-      );
-    })
-  );
-
 export default [
   changeTemplate,
-  editTemplate,
-  duplicateItem,
-  duplicateAsset
+  editTemplate
 ] as Epic[];
